move array specific functions into own module
